cburk2022
I am interested in developing a security app for the enterprise. I am having to learn a pile of things on the fly such as VB, K8S, Docker and MySql. I have developed apps for Android and IOS along with websites but its been a minute since I sold my last app.